chore: separate set of players from JetMinecraftServer to PlayerList
